+++ composeApp/src/iosMain/kotlin/com/focusclinic/app/di/PlatformModule.ios.kt
@@ -1,0 +1,9 @@
+package com.focusclinic.app.di
+
+import com.focusclinic.data.database.DriverFactory
+import org.koin.core.module.Module
+import org.koin.dsl.module
+
+actual fun platformModule(): Module = module {
+    single { DriverFactory() }
+}